Use a separate identity signing key for x3dh
This deviates from the Signal protocol, which opts to reuse the identity
X25519 Diffie-Hellman key as an Ed25519 signing key to sign ephemeral
keys. Instead of implementing a sketchy birational equivalence and trusting
users to not sign arbitrary messages, I decided to just extract the signing
functionality of the identity key and close that can of worms.